Move EntityBase back to common
